+++ src/transclude.ts
@@ -260,12 +260,14 @@ export abstract class TranscludeToken extends Token {
 
 	/** 获取模板或模块名 */
 	#getTitle(): Title {
-		const isTemplate = this.type === 'template';
-		return this.normalizeTitle(
-			this.childNodes[isTemplate ? 0 : 1].toString(true),
-			isTemplate ? 10 : 828,
-			true,
-		);
+		const isTemplate = this.type === 'template',
+			title = this.normalizeTitle(
+				this.childNodes[isTemplate ? 0 : 1].toString(true),
+				isTemplate ? 10 : 828,
+				true,
+			);
+		title.fragment = undefined;
+		return title;
 	}
 
 	/**