+++ base.ts
@@ -340,6 +340,7 @@ export type CompletionItem = Omit<CompletionItemBase, 'kind'> & {
 };
 
 export interface LanguageService {
+	include: boolean;
 	/** @private */
 	data?: SignatureData;