+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [`Token.prototype.buildLists`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token#buildlists) that builds all `list-range` tokens
 - [`LanguageService.prototype.provideDiagnostics`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService#providediagnostics) now provides diagnostics for inline CSS of extension tags, HTML tags and tables
 - [`LanguageService.prototype.provideDiagnostics`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService#providediagnostics) now provides diagnostics for embedded JSON in extension tags (`templatedata`, `mapframe`, `maplink`, and `graph`)
+- [`LanguageService.prototype.provideHover`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService#providehover) now provides hover for inline CSS of extension tags, HTML tags and tables
 
 **Fixed**