+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - [`Token.prototype.findEnclosingHtml`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token#findenclosinghtml)
 - [`RedirectToken.prototype.cloneNode`](https://github.com/bhsd-harry/wikiparser-node/wiki/RedirectToken#clonenode)
+- [`LanguageService.prototype.provideCompletionItems`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService#providecompletionitems) for attribute keys of extension tags, HTML tags and tables
 
 **Changed**